move: kafkaconfig 패키지로 이동
