+++ src/main/java/com/skyhorsemanpower/BE_AuctionPost/config/kafkaconfig/dto/InitialAuctionDto.java
@@ -1,15 +1,11 @@
-package com.skyhorsemanpower.BE_AuctionPost.kafka.dto;
+package com.skyhorsemanpower.BE_AuctionPost.config.kafkaconfig.dto;
 
-import com.skyhorsemanpower.BE_AuctionPost.data.dto.CreateAuctionPostDto;
-import com.skyhorsemanpower.BE_AuctionPost.status.AuctionEndTimeState;
-import com.skyhorsemanpower.BE_AuctionPost.status.TimeZoneChangeEnum;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
 
 @Getter
 @ToString
